add separate folder for authentication and users table
