+++ app/messages/websocket.py
@@ -1,7 +1,7 @@
 
 from fastapi import WebSocket, WebSocketDisconnect, APIRouter, Depends, HTTPException
 from ..websocket_manager import ConnectionManager
-from ..users.auth import get_current_user, websocket_authentication
+from ..authentication.services import websocket_authentication
 from .crud import insert_messages_to_db, insert_text_messages_to_db, get_sender_data
 from ..db_instance import db
 from ..utils.logger import logger
@@ -25,7 +25,7 @@ async def websocket_message_endpoint(websocket: WebSocket, workspace_id: int, us
             data = await websocket.receive_text()
             payload = json.loads(data)
             sender_cache = {}
-            
+
             message_data = {
                 "workspace_id": workspace_id,
                 "sender_id": user_id,
@@ -46,8 +46,6 @@ async def websocket_message_endpoint(websocket: WebSocket, workspace_id: int, us
                     if payload["type"] == "text":
                         await insert_text_messages_to_db(text_data={"message_id": message_id, "content":payload["content"]}, conn=conn)
                         
-
-
             message_obj = {
                 "message_id": message_id,
                 "workspace_id": workspace_id,
