+++ app/task/routes/task_router.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from typing import List
 import asyncpg
 from ...deps import get_db_connection  
-from ...users.auth import get_current_user
+from ...authentication.services import get_current_user
 from ...task.schemas.Tasks_schema import TaskCreate, TaskPatch, TaskAllOut
 from ...task.crud import task_crud
 from ...utils.decorators import db_error_handler
