+++ app/users/schemas.py
@@ -1,17 +1,6 @@
 from pydantic import BaseModel, AwareDatetime
 from typing import Optional
 
-
-class UserRegisterSchema(BaseModel):
-    first_name: str
-    last_name: str
-    email: str
-    password_hash: str
-    
-class UserLoginSchema(BaseModel):
-    email: str
-    password_hash: str
-
 class UserBasic(BaseModel):
     user_id: int
     profile_picture: Optional[str] = None
@@ -28,11 +17,6 @@ class UserDetail(BaseModel):
     profile_picture: Optional[str] = None
     status: Optional[str] = None
     phone_number: Optional[str] = None
-class RefreshToken(BaseModel):
-    refresh_token: str    
-
-class EmailObject(BaseModel):
-    email: str
     
 class GetUserResponse(BaseModel):
     first_name:str
